Refactor: 자동스크롤 관련 로직을 커스텀 훅으로 분리, EditorPanel의 응집도 강화
